Put files into packages
