+++ src/main/java/duke/tasks/Events.java
@@ -1,4 +1,7 @@
-import java.time.LocalDate;
+package duke.tasks;
+
+import duke.ui.GengException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
