+++ src/main/java/duke/tasks/ToDos.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public class ToDos extends Task{
     public ToDos(String description) {
         super(description);
