+++ src/main/java/duke/ui/GengException.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 public class GengException extends Exception {
     public GengException(String message) {
         super(message);
