+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,10 @@
+package duke.ui;
+
+import duke.tasks.Deadlines;
+import duke.tasks.Events;
+import duke.tasks.Task;
+import duke.ui.GengException;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
